requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.4.3
Django==5.0.2
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
filelock==3.13.1
gdown==5.1.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
Markdown==3.5.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pluggy==1.4.0
psycopg2==2.9.9
PySocks==1.7.1
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-mock==1.11.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
